Got the altimeter working in the flight software. The ALT_test() still doesn't fuction, but I AM getting altitude data. For some reason the first telemetry pull doesn't get the altitude data, and everything is just zeros. I've looked into the reason but don't have any idea why that's happeneing. Luckily, this doesn't affect the rest of the code, since the sensor::calibrate() function calls for the data five times before allowing the code to actually begin. :)

+++ FlightSW_test/sensor.h
@@ -12,7 +12,7 @@ class sensor {
     int16_t gyro_x;
     int16_t gyro_y;
     int16_t gyro_z;
-    int16_t altitude;
+    int32_t altitude;
 
     int16_t accel_x_bias;
     int16_t accel_y_bias;
